Refactored Weight
